+++ src/screens/VoteList/List.js
@@ -1,9 +1,14 @@
 /* eslint no-underscore-dangle: ["error", { "allow": ["_id"] }] */
 import React, { Component } from "react";
+import {
+  TouchableHighlight,
+  Dimensions,
+  Platform,
+  ActivityIndicator
+} from "react-native";
 import styled from "styled-components/native";
 import PropTypes from "prop-types";
 import { Navigator } from "react-native-navigation";
-import { Dimensions, Platform } from "react-native";
 import { graphql } from "react-apollo";
 import { unionBy } from "lodash";
 import Ionicons from "react-native-vector-icons/Ionicons";
@@ -19,6 +24,12 @@ const Wrapper = styled.View`
   flex: 1;
   background-color: #fff;
   width: ${({ width }) => width};
+`;
+
+const Loading = styled.View`
+  height: 50;
+  align-items: center;
+  justify-content: center;
 `;
 
 const SectionList = styled.SectionList``;
@@ -136,6 +147,13 @@ class List extends Component {
     return (
       <Wrapper onLayout={this.onLayout} width={this.state.width}>
         <SectionList
+          ListFooterComponent={() =>
+            data.loading ? (
+              <Loading>
+                <ActivityIndicator />
+              </Loading>
+            ) : null
+          }
           sections={this.prepareData()}
           stickySectionHeadersEnabled
           keyExtractor={({ _id }) => _id}