+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add time until goverment-vote [#358](https://github.com/demokratie-live/democracy-client/issues/358)
 - [Details] History segment of current states to Detail page
 - [iOS] network indicator in status bar [#367](https://github.com/demokratie-live/democracy-client/issues/367)
+- [ListView] load more indicator [#368](https://github.com/demokratie-live/democracy-client/issues/368)
 
 ### Changed
 - [ListView] improve performance by shouldUpdateComponent check [#362](https://github.com/demokratie-live/democracy-client/issues/362)